Fixed GPS position timestamp data type

+++ src/main/java/de/tuberlin/amos/ws17/swit/gps/GpsPosition.java
@@ -22,7 +22,7 @@ public class GpsPosition {
 	 */
 	private double longitude;
 
-	public double getTimeStamp() {
+	public long getTimeStamp() {
 		return timeStamp;
 	}
